Fix building from source in test kitchen
Add build-essential cookbook to berkshelf and kitchen.yml so build
tools are available. This isn't a dependency of the cookbook in
metadata because the source recipe is optional and package
installation is generally recommended where available.

Remove the compile flags for Ubuntu as the defaults now work for
CouchDB 1.2.1.

Use libmozjs-dev (virtual package) on all Ubuntu versions but 10.04.